requirements.txt
appdirs==1.4.3
Django==2.1.1
djangorestframework==3.9.1
packaging==16.8
pyparsing==2.3.1
pytz==2018.9
six==1.12.0
